Refactored SelectionModeViewController to use new collection view
